fixing bugs WIP

+++ ft_format_str.c
@@ -6,7 +6,7 @@
 /*   By: tvanelst <tvanelst@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/22 15:24:21 by tvanelst          #+#    #+#             */
-/*   Updated: 2021/03/22 20:56:14 by tvanelst         ###   ########.fr       */
+/*   Updated: 2021/03/22 21:39:36 by tvanelst         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ char	*ft_format_str(char *s, int size)
 {
 	char	*str;
 
-	if (size < 0 || size > ft_strlen(s))
+	if (size < 0 || (unsigned int)size > ft_strlen(s))
 		size = ft_strlen(s);
 	str = malloc(size + 1);
 	ft_strlcpy(str, s, size + 1);
